+++ src/App.js
@@ -6,7 +6,7 @@ import Login from "./pages/Login/Login";
 import SignUp from "./pages/SignUp/SignUp";
 import Home from "./pages/Home/Home";
 import PrivatePage from "./components/privatePage/PrivatePage";
-import UserPage from "./components/userPage/userPosts";
+import UserPage from "./pages/userPage/userPosts";
 import HashtagsPage from "./pages/hashtagsPage/hashtagsPage";
 
 export default function App() {
