+++ src/assets/styles/Logo.js
@@ -1,8 +1,8 @@
 import styled from "styled-components";
 
-export default function Logo() {
+export default function Logo({onClick}) {
     return (
-        <Wrapper>
+        <Wrapper onClick={onClick}>
             linkr
         </Wrapper>
     );
@@ -13,6 +13,10 @@ const Wrapper = styled.h1`
     color: var(--main-white);
     font-size: 49px;
     font-weight: 700;
+
+    :hover {
+        cursor: pointer;
+    }
 
     @media (max-width: 614px) {
         font-size: 45px;
